modify: draw nutritionistTabs
